feat: the id (s) will now be set automatically without that you need to do anything.

+++ postgresql/src/main/java/com/example/postgresql/repository/ServiceRepository.java
@@ -8,6 +8,6 @@ import java.sql.Timestamp;
 
 
 @Repository
-public interface ServiceRepository extends JpaRepository<ServiceDataResponse, Integer>
+public interface ServiceRepository extends JpaRepository<ServiceDataResponse, Long>
 {
 }
